change naming to logLayer
