+++ frontend/entrypoints/main.js
@@ -1,4 +1,4 @@
-import "~/javascripts/log/log_data_layer"
+import "~/javascripts/logs/log_layer"
 import * as Turbo from "@hotwired/turbo"
 import { Application } from "stimulus"
